requirements.txt
appconfig==0.1
appdirs==1.4.4
apturl==0.5.2
asgiref==3.5.2
asttokens==2.0.5
backcall==0.2.0
backports.zoneinfo==0.2.1
bcrypt==3.1.7
blinker==1.4
Brlapi==0.7.0
certifi==2022.6.15
chardet==3.0.4
charset-normalizer==2.1.0
click==8.1.3
cloudinary==1.29.0
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
decorator==5.1.1
defer==1.0.6
distlib==0.3.5
distro==1.4.0
distro-info===0.23ubuntu1
dj-database-url==1.0.0
Django==4.1
django-cloudinary-storage==0.3.0
django-cors-headers==3.13.0
django-heroku==0.3.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
duplicity==0.8.12.0
entrypoints==0.3
executing==0.9.1
fasteners==0.14.1
filelock==3.8.0
future==0.18.2
gcd==0.1.0
gpg===1.13.1-unknown
greenlet==0.4.15
gunicorn==20.1.0
httplib2==0.14.0
idna==3.3
ipython==8.4.0
jedi==0.18.1
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
matplotlib-inline==0.1.3
monotonic==1.5
msgpack==0.6.2
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
paramiko==2.6.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pipenv==2022.8.17
platformdirs==2.5.2
prompt-toolkit==3.0.30
protobuf==3.6.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
Pygments==2.12.0
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pynvim==0.4.1
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.8
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
pytz==2022.2.1
pyxdg==0.26
PyYAML==5.3.1
ranger-fm==1.9.3
reportlab==3.5.34
requests==2.28.1
requests-unixsocket==0.2.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.16.0
sqlparse==0.4.2
stack-data==0.4.0
systemd-python==234
tk==0.1.0
traitlets==5.3.0
ubuntu-advantage-tools==27.9
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.11
usb-creator==0.3.7
virtualenv==20.16.3
virtualenv-clone==0.5.7
wadllib==1.3.3
wcwidth==0.2.5
whitenoise==6.2.0
xkit==0.0.0
